+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.11.0] - 2017-09-09
 ### Added
 - option to not show break ideas
 - option to show notification 30 seconds before break starts
@@ -150,7 +152,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - resume/pause functionality for reminder
 - scripts for creating installers for OS X, Windows, Linux
 
-[Unreleased]: https://github.com/hovancik/stretchly/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/hovancik/stretchly/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/hovancik/stretchly/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/hovancik/stretchly/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/hovancik/stretchly/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/hovancik/stretchly/compare/v0.8.0...v0.8.1